chatgpt_tool/requirements.txt
# Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# LLM
openai==1.3.8
langchain==0.1.0

# Vector Store
chromadb==0.4.17

# PDF Processing
unstructured==0.11.5
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.1.0
poppler-utils==0.1.0

# Data
pandas==2.1.3
numpy==1.26.2

# Utilities
requests==2.31.0
networkx==3.2
tqdm==4.66.1
tenacity==8.2.3

# Frontend
streamlit==1.29.0
plotly==5.18.0

# PDF Generation
reportlab==4.0.7